package, batch batch, cluster, sensor, cluster The category. It is added and imported and used in the harvest
