+++ harvests/templates/sensor.html
