+++ sources/views.py
@@ -9,15 +9,3 @@ def vehicles(request):
 
 def parcel(request):   
     return render(request, "parcel.html")
-
-def bunchbatch(request):   
-    return render(request, "bunchbatch.html")    
-
-def bunch(request):   
-    return render(request, "bunch.html")  
-
-def sensor(request):   
-    return render(request, "sensor.html")    
-
-def bunchcategory(request):   
-    return render(request, "bunchcategory.html")
